V. 0 (Boilerplate)
Added all base-code to have bootstrap ready with CDN assets included and some basic folders for css, js and images.

See readme.md for instructions for your SASS compiler if you are uncertain about what to compile and where to save the compiled output. 

Hint: You can also see in index.html (head) what css is included and WHERE it is supposed to be included from. That is your output path.